Move to pytest-only file
