+++ nose_unit.cfg
@@ -6,5 +6,5 @@ nologcapture=1
 verbosity=2
 where=yt
 with-timer=1
-ignore-files=(test_load_errors.py|test_load_sample.py|test_commons.py)
+ignore-files=(test_load_errors.py|test_load_sample.py|test_commons.py|test_field_access_pytest.py)
 exclude-test=yt.frontends.gdf.tests.test_outputs.TestGDF
