+++ tests/tests.yaml
@@ -181,6 +181,7 @@ other_tests:
      - '--ignore-files=test_load_errors.py'
      - '--ignore-files=test_commons.py'
      - '--ignore-files=test_load_sample.py'
+     - '--ignore-files=test_field_access_pytest.py'
      - '--exclude-test=yt.frontends.gdf.tests.test_outputs.TestGDF'
   cookbook:
      - 'doc/source/cookbook/tests/test_cookbook.py'
